requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
cryptography==46.0.3
python-multipart==0.0.21
orjson==3.11.5
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.0